requirements.txt
streamlit
Pillow
numpy
opencv-python-headless
streamlit-image-coordinates
streamlit-canvas
rembg[cpu]
onnxruntime
